Move registry functions in views to new registry.py file
